wot_server/requirements.txt
# WoT AI Training Server (Linux)
# 用于模型训练、强化学习和深度学习任务

# Deep Learning & RL
torch>=2.0.0
torchvision>=0.15.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
tensorboard>=2.12.0

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0  # YOLOv8
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
python-dotenv>=1.0.0

# API Integration (用于与客户端通信)
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.104.0  # REST API 服务器
uvicorn>=0.24.0   # ASGI 服务器
websockets>=12.0  # WebSocket 支持

# Development
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0

